feat(line-items): move extraction core into receipt_upload with a CI golden gate
Pure relocation of the extractor's logic plus the test coverage it never had:

- receipt_upload/line_items/geometry.py: estimate_skew, band_words,
  parse_band, extract_items, reconcile and their regexes, moved VERBATIM
  from scripts/extract_line_items.py. Verified as a pure relocation by
  running both versions over all 25 golden receipts' stored OCR:
  25/25 byte-identical item lists.
- scripts/extract_line_items.py becomes a CLI shim (fetch + print only);
  the hardcoded /Users/tnorlund sys.path hack is replaced with repo-root
  resolution. No extraction logic remains in scripts/.
- scripts/tests/test_extract_line_items.py -> receipt_upload/tests/
  test_line_item_geometry.py. These 12 tests were INVISIBLE to CI: the
  repository-tests job collects `find scripts -maxdepth 1`, which never
  descends into scripts/tests/. They now run on every PR.
- NEW golden regression gate: test_line_item_golden_regression.py runs
  extract_items over a hermetic OCR fixture (line_items_golden_ocr.json,
  word geometry captured from dev -- no AWS in CI) and enforces
  per-merchant recall/name floors pinned at the measured post-deskew
  baseline. Aggregate metrics are not enough: Trader Joe's 0% names hid
  inside a 35% average. Sensitivity proven by running the pre-deskew
  extractor through the gate -- it fails on exactly the two Trader Joe's
  receipts (names 0% and 29% against 100% floors).
- scripts/export_line_item_golden_ocr.py regenerates the OCR fixture when
  a golden receipt's sections legitimately change; the hand-labeled truth
  never regenerates.

Floors are floors, not targets: raise them when a merchant genuinely
improves; never lower them to make a change pass.
